编译告警整改

+++ sensor/hdi_service/sensor_clients_manager.cpp
@@ -213,7 +213,7 @@ void SensorClientsManager::UpdateClientPeriodCount(int sensorId, int64_t samplin
     for (auto &entry : clients_[groupId]) {
         auto &client = entry.second;
         if (client.curCountMap_.find(sensorId) == client.curCountMap_.end() ||
-            static_cast<uint32_t>(HdfRemoteGetCallingPid()) == entry.first) {
+            HdfRemoteGetCallingPid() == entry.first) {
             client.curCountMap_[sensorId] = INIT_CUR_COUNT;
         }
         if (client.sensorConfigMap_.find(sensorId) != client.sensorConfigMap_.end()) {
